# apps/api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/autenticacao/web/pareamentos-qr": {
      "post": {
        "operationId": "gerarPareamentoQrWeb",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PareamentoQrGeradoDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessaoWeb": []
          }
        ],
        "summary": "Gera um token QR efêmero vinculado à sessão web",
        "tags": [
          "autenticacao"
        ]
      }
    },
    "/api/v1/autenticacao/web/pareamentos-qr/{pareamentoId}": {
      "get": {
        "operationId": "consultarPareamentoQrWeb",
        "parameters": [
          {
            "name": "pareamentoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumoPareamentoQrWebDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessaoWeb": []
          }
        ],
        "summary": "Consulta o aparelho que aguarda confirmação nesta sessão web",
        "tags": [
          "autenticacao"
        ]
      }
    },
    "/api/v1/autenticacao/web/pareamentos-qr/{pareamentoId}/confirmar": {
      "post": {
        "operationId": "confirmarPareamentoQrWeb",
        "parameters": [
          {
            "name": "pareamentoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "sessaoWeb": []
          }
        ],
        "summary": "Confirma um aparelho com autenticação web recente",
        "tags": [
          "autenticacao"
        ]
      }
    },
    "/api/v1/autenticacao/web/pareamentos-qr/{pareamentoId}/cancelar": {
      "post": {
        "operationId": "cancelarPareamentoQrWeb",
        "parameters": [
          {
            "name": "pareamentoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "sessaoWeb": []
          }
        ],
        "summary": "Cancela o pareamento vinculado à sessão web atual",
        "tags": [
          "autenticacao"
        ]
      }
    },
    "/api/v1/autenticacao/web/entrar": {
      "post": {
        "operationId": "entrarSessaoWeb",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntradaLoginWebDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessaoWebDto"
                }
              }
            }
          }
        },
        "summary": "Autentica e cria uma sessão web",
        "tags": [
          "autenticacao"
        ]
      }
    },
    "/api/v1/autenticacao/web/sessoes": {
      "get": {
        "operationId": "listarSessoesWeb",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResumoSessaoWebDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "sessaoWeb": []
          }
        ],
        "summary": "Lista as sessões web ativas do usuário atual",
        "tags": [
          "autenticacao"
        ]
      }
    },
    "/api/v1/autenticacao/web/sessao": {
      "get": {
        "operationId": "obterSessaoWeb",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessaoWebDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessaoWeb": []
          }
        ],
        "summary": "Obtém a sessão web atual",
        "tags": [
          "autenticacao"
        ]
      }
    },
    "/api/v1/autenticacao/web/rotacionar": {
      "post": {
        "operationId": "rotacionarSessaoWeb",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessaoWebDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessaoWeb": []
          }
        ],
        "summary": "Rotaciona os segredos da sessão web",
        "tags": [
          "autenticacao"
        ]
      }
    },
    "/api/v1/autenticacao/web/sair": {
      "post": {
        "operationId": "sairSessaoWeb",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "sessaoWeb": []
          }
        ],
        "summary": "Revoga a sessão web atual",
        "tags": [
          "autenticacao"
        ]
      }
    },
    "/api/v1/autenticacao/web/sair-todas": {
      "post": {
        "operationId": "sairTodasSessoesWeb",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "sessaoWeb": []
          }
        ],
        "summary": "Revoga todas as sessões web do usuário atual",
        "tags": [
          "autenticacao"
        ]
      }
    },
    "/api/v1/autenticacao/web/sessoes/{sessaoId}/revogar": {
      "post": {
        "operationId": "revogarSessaoWebDoUsuario",
        "parameters": [
          {
            "name": "sessaoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "sessaoWeb": []
          }
        ],
        "summary": "Revoga uma sessão web pertencente ao usuário atual",
        "tags": [
          "autenticacao"
        ]
      }
    },
    "/api/v1/autenticacao/web/usuarios/{usuarioId}/revogar-sessoes": {
      "post": {
        "operationId": "revogarSessoesWebAdministrativamente",
        "parameters": [
          {
            "name": "usuarioId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "sessaoWeb": []
          }
        ],
        "summary": "Revoga administrativamente as sessões web de um usuário",
        "tags": [
          "autenticacao"
        ]
      }
    },
    "/api/v1/autenticacao/web/usuarios/{usuarioId}/revogar-dispositivos-mobile": {
      "post": {
        "operationId": "revogarDispositivosMobileAdministrativamente",
        "parameters": [
          {
            "name": "usuarioId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "sessaoWeb": []
          }
        ],
        "summary": "Revoga administrativamente os dispositivos móveis de um usuário",
        "tags": [
          "autenticacao"
        ]
      }
    }
  },
  "info": {
    "title": "API Vyntra Omnichannel",
    "description": "Contrato público e versionado do Omnichannel V1.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "sessaoWeb": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-vyntra_sessao"
      },
      "sessaoMobile": {
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "type": "http"
      }
    },
    "schemas": {
...
      "PareamentoQrGeradoDto": {
        "type": "object",
        "properties": {
          "pareamento_id": {
            "type": "string",
            "format": "uuid"
          },
          "token_qr": {
            "type": "string",
            "minLength": 43,
            "maxLength": 43
          },
          "expira_em": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "pareamento_id",
          "token_qr",
          "expira_em"
        ]
      },
      "ResumoPareamentoQrWebDto": {
        "type": "object",
        "properties": {
          "pareamento_id": {
            "type": "string",
            "format": "uuid"
          },
          "estado": {
            "type": "string",
            "enum": [
              "AGUARDANDO_RESGATE",
              "AGUARDANDO_CONFIRMACAO",
              "CONFIRMADO"
            ]
          },
          "expira_em": {
            "type": "string",
            "format": "date-time"
          },
          "plataforma": {
            "type": "string",
            "enum": [
              "IOS",
              "ANDROID"
            ]
          },
          "modelo_sanitizado": {
            "type": "string"
          },
          "versao_aplicativo": {
            "type": "string"
          }
        },
        "required": [
          "pareamento_id",
          "estado",
          "expira_em"
        ]
      },
      "EntradaLoginWebDto": {
        "type": "object",
        "properties": {
          "confirmar_revogacao_sessao_mais_antiga": {
            "type": "boolean",
            "default": false,
            "description": "Confirma a substituição da sessão web ativa mais antiga quando o limite for alcançado."
          },
          "identificador": {
            "type": "string",
            "example": "maria.silva",
            "maxLength": 120,
            "minLength": 3
          },
          "senha": {
            "type": "string",
            "example": "uma senha longa e privada",
            "maxLength": 128,
            "minLength": 12
          }
        },
        "required": [
          "identificador",
          "senha"
        ]
      },
      "SessaoWebDto": {
        "type": "object",
        "properties": {
          "sessao_id": {
            "type": "string",
            "format": "uuid"
          },
          "usuario_id": {
            "type": "string",
            "format": "uuid"
          },
          "nome_exibicao": {
            "type": "string",
            "example": "Maria Silva"
          },
          "expira_em": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "sessao_id",
          "usuario_id",
          "nome_exibicao",
          "expira_em"
        ]
      },
      "ResumoSessaoWebDto": {
        "type": "object",
        "properties": {
          "sessao_id": {
            "type": "string",
            "format": "uuid"
          },
          "atual": {
            "type": "boolean"
          },
          "autenticada_em": {
            "type": "string",
            "format": "date-time"
          },
          "ultima_atividade_em": {
            "type": "string",
            "format": "date-time"
          },
          "expira_em": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "sessao_id",
          "atual",
          "autenticada_em",
          "ultima_atividade_em",
          "expira_em"
        ]
      }
    }
  }
}